WFDM-1185 Remove the boxing to "Double".

+++ wfdm-file-index-service/src/main/java/ca/bc/gov/nrs/wfdm/wfdm_file_index_service/OpenSearchRESTClient.java
@@ -381,7 +381,7 @@ public class OpenSearchRESTClient {
 			factor = 1000;
 		}
 
-		return Double.valueOf(ret * Math.pow(factor, power)).longValue();
+		return (long) (ret * Math.pow(factor, power));
 	}
 	
 }
